Move pick operator test of CEDDL over to the pick operator spec.
